Adding buttons, changing styles
In this I added some buttons and changed some positions of the elements
on the screen.  I also hid the default file chooser and linked it to a
button instead.  I didn't know what color i should have gone with for
the web controls so i just used red, Cheers!